+++ src/map/map.cpp
@@ -1207,7 +1207,7 @@ int32 map_config_read(const int8* cfgName)
         }
         else if (strcmp(w1,"engage_delay_mod") == 0)
         {
-            map_config.engage_delay_mod = atoi(w2);
+            map_config.engage_delay_mod = (float)atof(w2);
         }
         else if (strcmp(w1, "speed_mod") == 0)
         {
